comments/requirements.txt
click==8.0.1
Flask==2.0.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
grpcio==1.41.0
grpcio-tools==1.30.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
protobuf==3.18.0
pymssql==2.2.2
python-dotenv==0.19.0
six==1.16.0
Werkzeug==2.0.1
WTForms==2.3.3
